// templates/workday/types.ts
interface Value {
    attributes: {
        [key: string]: string;
        'wd:type': string;
    };
    $value: string;
}
// ...
export interface ResponseWorkdayLocation {
    Location_Reference: {
        ID: Value[];
    };
// ...
}